twitter_data/requirements.txt
requests
#requests_oauthlib>=1.3.0
minio>=7.1.2
python-dotenv>=0.19.2
